Bash in truncate
Truncate is completely wrong. Just smashed it in to get it done so
as to explore the problem domain; which is that there is no
topology graph! Conway polyhedral operations are done on polyhedron
topologies. The actual 3D position of vertices is entirely up to the
implementer if they so want to render the polyhedron. I was doing
this completely wrong.

HUGE refactor incoming. I also need to split out the GPU stuff.